requirements.txt
torch==2.4.0
accelerate==0.34.0
opencv-python==4.10.0.84
pillow==10.4.0
numpy==1.26.4
timm==1.0.11
kornia==0.7.1
scikit-image==0.24.0
huggingface-hub==0.24.6
transformers==4.46.3
gradio==4.26.0
fastapi==0.109.2
